separate nodes test from the suite test
